requirements/default.txt
django-picklefield
celery>=2.5.0
